Refactoring: Create PageReader class and move page reading logic here Move getBookDetailsLinkInformation and getContributorsStringWithSpecialSeparationCharacter methods here and create getBookInformation method
