+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/Crawler.java
@@ -25,11 +25,10 @@ public class Crawler implements BookCrawler {
 
     private final BookCreator bookCreator;
     private final URLFactory urlFactory;
+    private final PageReader pageReader;
 
     private static final String ISBN963 = "978963";
     private static final String ISBN615 = "978615";
-    private static final String SPECIAL_CASE_OTHER_NAMES = "Egyéb nevek:";
-    private static final String SPECIAL_SEPARATION_CHARACTER = "$";
     private static final String PAGE_SIZE = "10";
 
     private int page = 0;
@@ -81,55 +80,20 @@ public class Crawler implements BookCrawler {
         try {
             final Document bookListPage = Jsoup.connect(url).get();
 
-            return getBookDetailsLinkInformation(bookListPage);
+            return pageReader.getBookDetailsLinkInformation(bookListPage);
         } catch (IOException e) {
             throw new BookListDownloadException("Unable to download the list book page! Url: " + url, e);
         }
     }
 
-    private Map<String, String> getBookDetailsLinkInformation(Document document) {
-        Map<String, String> booksInformation = new HashMap<>();
-
-        Elements informationLines = document.select(".short_item_back script");
-        for (Element line : informationLines) {
-            Pattern lineNumberAndBookId = Pattern.compile("(\\d+,\\d+)");
-            Matcher matcher = lineNumberAndBookId.matcher(line.toString());
-            if (matcher.find()) {
-                String[] information = matcher.group().split(",");
-                int lineNumber = 0;
-                int bookId = 1;
-                booksInformation.put(information[lineNumber], information[bookId]);
-            }
-        }
-        return booksInformation;
-    }
-
     private Optional<Book> getBook(String bookDetailsUrl) {
         try {
             Document bookPage = Jsoup.connect(bookDetailsUrl).get();
-            Elements bookProperties = bookPage.select(".long_key");
-            Elements bookPropertiesValues = bookPage.select(".long_value");
-
-            Map<String, String> prepareBook = new HashMap<>();
-            for (int i = 0; i < bookProperties.size() - 1; i++) {
-                if (bookProperties.get(i).text().equals(SPECIAL_CASE_OTHER_NAMES) && !bookPropertiesValues.get(i).select("a").isEmpty()) {
-                    Elements names = bookPropertiesValues.get(i).select("a");
-                    String contributorsName = getContributorsStringWithSpecialSeparationCharacter(names);
-                    prepareBook.put(bookProperties.get(i).text().trim(), contributorsName);
-                    break;
-                }
-                prepareBook.put(bookProperties.get(i).text().trim(), bookPropertiesValues.get(i).text().trim());
-            }
-
-            return bookCreator.createBook(prepareBook, SPECIAL_SEPARATION_CHARACTER);
+            return bookCreator.createBook(
+                    pageReader.getBookInformation(bookPage),
+                    PageReader.getSpecialSeparationCharacter());
         } catch (IOException e) {
             throw new BookDownloadException("Unable to download book page! Url: " + bookDetailsUrl, e);
         }
-    }
-
-    private String getContributorsStringWithSpecialSeparationCharacter(Elements names) {
-        Optional<String> otherNames = names.stream().map(Element::text).reduce((s, s2) -> s + SPECIAL_SEPARATION_CHARACTER + s2);
-        String otherNamesArraysToString = Arrays.toString(otherNames.stream().toArray());
-        return otherNamesArraysToString.substring(1, otherNamesArraysToString.length() - 1);
     }
 }
